(#15717) Kick agent using its fqdn
Previously, the acceptance test would fail if `puppet kick` verified the
SSL certificate of the target, and the hostname it connected to didn't
match what was in the certificate.

The `agent` string contains the agent hostname as specified in the
configuration file, not the fqdn. Also, the agent and master may be in
different domains.

This commit modifies the test to query the agent for its fqdn, and then
puppet kick's that.