(#16757) Don't require rubygems
Previously, the windows service code was explicitly requiring
rubygems. While not breaking anything, it's not desirable, as the
environment should load rubygems, or not, instead of the service code
itself.

This code changes the daemon batch file to execute rubyw with rubygems
support enabled.

Paired-with: Andy Parker <andy@puppetlabs.com>