(#16922) Quote strings that contain ":"
Part of the performance work on the ZAML code caused a regression on how
to convert values that contain a colon into YAML strings. This becomes a
problem when a string ends in a colon and that line in the YAML output
is followed by another line.

An example of the incorrect output is:

  ---
  a: a:
  b: 1

This patch causes that to be output correctly as:

  ---
  a: "a:"
  b: 1