(#17000) Use --no-client argument to prevent agent from starting
Previously, the acceptance test would start a daemonized and listening
agent, and then attempt to kick it from the master. However, when the
agent started, it would automatically retrieve and apply a catalog. If
kicked during this time, the agent would return exit code 3, which is
ambiguous, since it could mean we're not authorized to kick the agent,
an SSL error prevented us from connecting, the agent couldn't find the
`run` indirection (which this acceptance test is attempting to ensure
never happens), or who knows what else.

In the case that the agent was 'running', the test would attempt to
retry the kick, but on centos55, the test would get into an infinite
loop.

This commit uses the `--no-client` agent argument to ensure the agent
only applies catalogs when it's kicked. In doing so, all of the retry
logic in the acceptance test can be removed.